Offline-warning, Pie Charts, delete Post
Warning when user is going offline -> jQuery Plugin „offline“
Pie Charts are now feed with the real data
When a user deletes a public gear it only changes the author id to 1
instead of deleting the post.

+++ library/register-new-gear.php
@@ -100,7 +100,15 @@ function registerNewGear(){
 		$deltePost = $_GET["deletepost"];
 		$deletePostId = $_GET["deletepostid"];
 		if ($deltePost != '') {
-			wp_delete_post($deletePostId);
+			if (get_post_type($deletePostId) == 'gear' && get_post_meta($deletePostId, 'gearpool', true)) {
+				$my_post = array(
+				  'ID'           => $deletePostId,
+				  'post_author' => 1
+				);
+			  	wp_update_post( $my_post );
+			}else{
+				wp_delete_post($deletePostId);				
+			}
 		}
 	}
